1047. Remove All Adjacent Duplicates In String
